# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Taiwan Tourism Open Data API
  description: '

    以 **交通部觀光署「景點 - 觀光資訊資料庫」** 政府開放資料（6,000+ 筆全台景點）為基礎的 RESTful API。


    資料來源：<https://data.gov.tw/dataset/7777>（觀光資料標準 V2.1）


    ### 資源 (Resources)

    | Resource | 說明 |

    |---|---|

    | `attractions` | 景點（支援關鍵字搜尋、多條件篩選、排序、分頁、附近景點查詢） |

    | `attractions/{id}/images` | 景點圖片（子資源） |

    | `categories` | 景點類型（文化類、生態類…） |

    | `cities` / `towns` | 縣市 / 鄉鎮市區 |

    | `reviews` | 遊客評論（1~5 星） |

    | `trips` / `trips/{id}/items` | 旅遊行程與行程項目 |

    | `stats` | 統計資訊（唯讀） |


    ### 驗證

    所有 **讀取 (GET)** 操作公開；**新增 / 修改 / 刪除 (POST / PUT / PATCH / DELETE)** 需在 Header
    帶入

    `X-API-Key: ntub-iot-2026`（可按右上角 **Authorize** 按鈕設定）。


    ### 慣例

    * 列表回應格式：`{"total", "page", "page_size", "pages", "items"}`，並附 `X-Total-Count`
    header

    * 建立成功回傳 `201 Created` 與 `Location` header；刪除成功回傳 `204 No Content`

    * 錯誤格式：`{"detail": "錯誤說明"}`（404 查無資料、409 資料衝突、422 驗證失敗）

    '
  contact:
    name: NTUB IoT App 作業
  license:
    name: 政府資料開放授權條款－第1版
    url: https://data.gov.tw/license
  version: 1.0.0
paths:
...
  /api/v1/categories:
    get:
      tags:
      - Categories 景點類型
      summary: 列出景點類型
      operationId: list_categories_api_v1_categories_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 名稱關鍵字
          title: Q
        description: 名稱關鍵字
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 頁碼（從 1 開始）
          default: 1
          title: Page
        description: 頁碼（從 1 開始）
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 每頁筆數（1~100）
          default: 20
          title: Page Size
        description: 每頁筆數（1~100）
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_CategoryOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Categories 景點類型
      summary: 新增景點類型
      operationId: create_category_api_v1_categories_post
      security:
      - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryOut'
        '409':
          description: 代碼或名稱重複
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/categories/{category_id}:
    get:
      tags:
      - Categories 景點類型
      summary: 取得單一景點類型
      operationId: get_category_api_v1_categories__category_id__get
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer
          title: Category Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Categories 景點類型
      summary: 整筆更新景點類型
      operationId: replace_category_api_v1_categories__category_id__put
      security:
      - APIKeyHeader: []
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer
          title: Category Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryBase'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Categories 景點類型
      summary: 部分更新景點類型
      operationId: update_category_api_v1_categories__category_id__patch
      security:
      - APIKeyHeader: []
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer
          title: Category Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Categories 景點類型
      summary: 刪除景點類型（景點本身不會被刪除，只移除關聯）
      operationId: delete_category_api_v1_categories__category_id__delete
      security:
      - APIKeyHeader: []
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer
          title: Category Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CategoryBase:
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 1
          title: Name
          examples:
          - 夜景類
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
          examples:
          - 適合欣賞夜景的地點
      type: object
      required:
      - name
      title: CategoryBase
    CategoryCreate:
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 1
          title: Name
          examples:
          - 夜景類
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
          examples:
          - 適合欣賞夜景的地點
        id:
          anyOf:
          - type: integer
            maximum: 9999.0
            minimum: 1.0
          - type: 'null'
          title: Id
          description: 類型代碼；省略時自動編號
          examples:
          - 100
      type: object
      required:
      - name
      title: CategoryCreate
    CategoryOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        attraction_count:
          type: integer
          title: Attraction Count
          default: 0
      type: object
      required:
      - id
      - name
      title: CategoryOut
...
    CategoryUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 50
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
      type: object
      title: CategoryUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Message:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: Message
...
    Page_CategoryOut_:
      properties:
        total:
          type: integer
          title: Total
          description: 符合條件的總筆數
        page:
          type: integer
          title: Page
          description: 目前頁碼（從 1 開始）
        page_size:
          type: integer
          title: Page Size
          description: 每頁筆數
        pages:
          type: integer
          title: Pages
          description: 總頁數
        items:
          items:
            $ref: '#/components/schemas/CategoryOut'
          type: array
          title: Items
      type: object
      required:
      - total
      - page
      - page_size
      - pages
      - items
      title: Page[CategoryOut]
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      description: 寫入操作需帶入 API Key（預設為 `ntub-iot-2026`，可用環境變數 `API_KEY` 修改）
      in: header
      name: X-API-Key
...
servers:
- url: http://127.0.0.1:8000
  description: 本機開發伺服器